move progress to it's own file
